rename settings scenes
