+++ src/components/@settings/@name/NameScene.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { object } from 'prop-types'
 import { withStyles, Typography, TextField, Button } from '@material-ui/core'
-import Helper from 'components/@settings/Helper'
+import Helper from '../Helper'
 import connector from '../connector'
 
 const styles = theme => ({
@@ -16,24 +16,24 @@ const styles = theme => ({
   },
 })
 
-const EmailScene = ({ classes, user }) =>
+const NameScene = ({ classes, user }) =>
   <div className={classes.root}>
     <div className={classes.input}>
-      <Typography variant="subheading">Email</Typography>
+      <Typography variant="subheading">Имя и фамилия</Typography>
       <TextField
         fullWidth
-        name="email"
-        placeholder="Email"
-        defaultValue={user.email}
+        name="name"
+        placeholder="Имя и фамилия"
+        defaultValue={user.name}
       />
     </div>
-    <Helper>Ваша почта будет видна всем всем людям на вашей вечеринке</Helper>
+    <Helper>Ваше имя и фамилия будут видны всем пользователям</Helper>
     <Button variant="raised" color="primary" className={classes.button}> Сохранить </Button>
   </div>
 
-EmailScene.propTypes = {
+NameScene.propTypes = {
   classes: object.isRequired,
   user: object.isRequired,
 }
 
-export default withStyles(styles)(connector(EmailScene))
+export default withStyles(styles)(connector(NameScene))
