+++ src/components/@settings/Helper.jsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import { object, node } from 'prop-types'
-import { withStyles } from '@material-ui/core/styles'
-import { Typography } from '@material-ui/core'
+import { withStyles, Typography } from '@material-ui/core'
 
 const styles = {
   root: {
