+++ src/components/@settings/SettingsLayout.jsx
@@ -3,11 +3,11 @@ import { object } from 'prop-types'
 import { Route, Switch } from 'react-router-dom'
 import { withStyles } from '@material-ui/core'
 import SettingsScene from './SettingsScene'
-import PasswordScene from './PasswordScene'
-import NameScene from './NameScene'
-import EmailScene from './EmailScene'
-import PhoneScene from './PhoneScene'
-import AvatarScene from './AvatarScene'
+import PasswordScene from './@password/PasswordScene'
+import NameScene from './@name/NameScene'
+import EmailScene from './@email/EmailScene'
+import PhoneScene from './@phone/PhoneScene'
+import AvatarScene from './@avatar/AvatarScene'
 import connector from './connector'
 
 const styles = {
